feat: PopulationGraphコンポーネントの作成
