+++ src/App.tsx
@@ -1,14 +1,10 @@
 import "./App.css"
 import { useState, useEffect } from "react"
 import getPrefData from "./components/prefAPI"
-import Highcharts from "highcharts"
-import HighchartsReact from "highcharts-react-official"
-import NoDataToDisplay from "highcharts/modules/no-data-to-display"
 import { makeNewStates } from "./makeNewStates"
-import { PrefData } from "./types"
+import { PopulationGraphData, PrefData } from "./types"
 import { PrefectureCheckbox } from "./components/PrefectureCheckbox/"
-
-NoDataToDisplay(Highcharts)
+import { PopulationGraph } from "./components/PopulationGraph"
 
 const App: React.FC = () => {
   const [prefAry, setPrefAry] = useState<PrefData[]>([])
@@ -17,85 +13,13 @@ const App: React.FC = () => {
     new Map<number, number[]>(),
   )
 
-  const graphData: { data: number[]; name: string }[] = checkedPrefCodes
+  const graphData: PopulationGraphData[] = checkedPrefCodes
     .map((code) => prefAry.find((pref) => pref.prefCode === code))
     .filter((pref) => pref !== undefined && loadedPrefData.has(pref.prefCode))
     .map((pref) => ({
       name: pref!.prefName,
       data: [...loadedPrefData.get(pref!.prefCode)!],
     }))
-
-  // 伸び代: グラフのコンポーネントをラップして別コンポーネントにしてしまいたい
-  const options = {
-    chart: {
-      type: "spline",
-      backgroundColor: "#fff",
-      polar: true,
-    },
-    title: {
-      text: "人口構成",
-    },
-    lang: {
-      noData: "データがありません",
-    },
-    noData: {
-      style: {
-        fontWeight: "bold",
-        fontSize: "15px",
-        color: "#303030",
-      },
-    },
-    xAxis: {
-      /* APIが返してくる年度はこれだった */
-      categories: [
-        "1960",
-        "1965",
-        "1970",
-        "1975",
-        "1980",
-        "1985",
-        "1990",
-        "1995",
-        "2000",
-        "2005",
-        "2010",
-        "2015",
-        "2020",
-        "2025",
-        "2030",
-        "2035",
-        "2040",
-        "2045",
-      ],
-      offset: 0,
-      title: {
-        text: "年度",
-        align: "high",
-        textAlign: "left",
-        rotation: 0,
-        offset: 0,
-        margin: 0,
-        y: 30,
-        x: -27,
-      },
-    },
-    yAxis: {
-      visible: true,
-      tickPosition: "inside",
-      offset: 0,
-      title: {
-        text: "人口数",
-        align: "high",
-        textAlign: "left",
-        rotation: 0,
-        offset: 0,
-        margin: 0,
-        y: -20,
-        x: -47,
-      },
-    },
-    series: graphData,
-  }
 
   useEffect(() => {
     getPrefData.GetPref().then((data) => setPrefAry(data))
@@ -129,10 +53,8 @@ const App: React.FC = () => {
       </div>
       <div className='container-chart'>
         {/* 伸び代: ワーニングがコンソールに出ている */}
-        <HighchartsReact
-          highcharts={Highcharts}
-          constructorType={"chart"}
-          options={options}
+        <PopulationGraph
+          graphData={graphData}
         />
       </div>
     </div>
