+++ src/types/index.ts
@@ -2,3 +2,5 @@ export type PrefData = {
   prefCode: number
   prefName: string
 }
+
+export type PopulationGraphData = { data: number[]; name: string };
